requirements.txt
annotated-types==0.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.2
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
MouseInfo==0.1.3
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.21.12
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.7.1
pydantic_core==2.18.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.2.0
pywin32==306
requests==2.32.2
rsa==4.9
SpeechRecognition==3.10.4
tqdm==4.66.4
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
